kyc_ai/requirements.txt
insightface==0.7.3
onnxruntime==1.18.0
opencv-python==4.10.0.84
numpy==1.26.4
pillow==10.4.0
pytesseract==0.3.13
